fix deserializacion cadena a uint8

+++ utils/src/utils/protocolo.h
@@ -52,7 +52,7 @@ typedef struct
 typedef struct
 {
     char *cadena;   
-    uint32_t entero; 
+    uint8_t entero; 
 } t_message_example;
 /*
  *  package functions
